Add probe-and-safe
Add `safe` parameter to `/probe` endpoint to allow battery-powered
devices to probe themselves, and save to probe results to be included
in the next `/metrics` call. This makes it possible to have up-to-date
metrics on battery-powered devices that are in sleep mode most of the
time and only wake up to push updates and will probably be sleeping when
`/metrics` is called.